feat(http): 实现内容协商和HTTP请求处理
- 新增 ContentNegotiator 类，用于协商最佳响应表示形式
- 添加 HttpProcessor 类，处理 HTTP 请求并生成响应
- 更新 WebServer 类，集成内容协商功能
- 优化 MultiLevelMemoryPool 类，提高内存分配效率
- 调整 benchmark 和 test 文件，适应新的逻辑

+++ benchmark/thread_pool_benchmark.cpp
@@ -14,7 +14,7 @@ static LoggerConfig loggerConfig;
 
 // 测试线程池任务提交性能
 static void BM_ThreadPoolSubmit(benchmark::State& state) {
-    ThreadPool pool(state.range(0));
+    ThreadPool pool(static_cast<size_t>(state.range(0)));
     for (auto _ : state) {
         pool.enqueue([](){});
     }
@@ -26,7 +26,7 @@ BENCHMARK(BM_ThreadPoolSubmit)
 
 // 测试线程池任务执行吞吐量
 static void BM_ThreadPoolThroughput(benchmark::State& state) {
-    ThreadPool pool(state.range(0));
+    ThreadPool pool(static_cast<size_t>(state.range(0)));
     std::atomic<int> counter{0};
     for (auto _ : state) {
         for (int i = 0; i < state.range(1); ++i) {
